+++ src/index.ts
@@ -14,7 +14,7 @@ export {
   InvalidExpressionError,
 } from './errors'
 export {
-  Database,
+  Schema,
   Table,
   BooleanColumn,
   NumberColumn,