+++ test/diffOptions.test.ts
@@ -1,4 +1,4 @@
-import { Builder, Database, Table, TextColumn } from '../src'
+import { Builder, Schema, Table, TextColumn } from '../src'
 
 describe('test Options', () => {
   const column1 = new TextColumn('col1')
@@ -8,12 +8,12 @@ describe('test Options', () => {
     'testTable',
     [column1, column2],
   )
-  const db = new Database([table], 1)
+  const schema = new Schema([table])
 
   describe('test Semicolon Option', () => {
-    const sqlWithoutSemicolon = new Builder(db, { useSemicolonAtTheEnd: false })
-    const sqlWithSemicolon = new Builder(db, { useSemicolonAtTheEnd: true })
-    const sqlDefault = new Builder(db)
+    const sqlWithoutSemicolon = new Builder(schema, { useSemicolonAtTheEnd: false })
+    const sqlWithSemicolon = new Builder(schema, { useSemicolonAtTheEnd: true })
+    const sqlDefault = new Builder(schema)
     it('Produces [SELECT 1 FROM "testTable"] without semicolon', () => {
       const actual = sqlWithoutSemicolon
         .select(1)
@@ -43,10 +43,10 @@ describe('test Options', () => {
   })
 
   describe('test OrderBy ASC Option', () => {
-    const sqlAlways = new Builder(db, { addAscAfterOrderByItem: 'always' })
-    const sqlNever = new Builder(db, { addAscAfterOrderByItem: 'never' })
-    const sqlWhenMentioned = new Builder(db, { addAscAfterOrderByItem: 'when mentioned' })
-    const sqlDefault = new Builder(db)
+    const sqlAlways = new Builder(schema, { addAscAfterOrderByItem: 'always' })
+    const sqlNever = new Builder(schema, { addAscAfterOrderByItem: 'never' })
+    const sqlWhenMentioned = new Builder(schema, { addAscAfterOrderByItem: 'when mentioned' })
+    const sqlDefault = new Builder(schema)
     it('Produces [SELECT "col1" FROM "testTable" ORDER BY "col1" ASC;] option(always)', () => {
       const actual = sqlAlways
         .select(column1)
@@ -129,10 +129,10 @@ describe('test Options', () => {
   })
 
   describe('test OrderBy NULLS LAST Option', () => {
-    const sqlAlways = new Builder(db, { addNullsLastAfterOrderByItem: 'always' })
-    const sqlNever = new Builder(db, { addNullsLastAfterOrderByItem: 'never' })
-    const sqlWhenMentioned = new Builder(db, { addNullsLastAfterOrderByItem: 'when mentioned' })
-    const sqlDefault = new Builder(db)
+    const sqlAlways = new Builder(schema, { addNullsLastAfterOrderByItem: 'always' })
+    const sqlNever = new Builder(schema, { addNullsLastAfterOrderByItem: 'never' })
+    const sqlWhenMentioned = new Builder(schema, { addNullsLastAfterOrderByItem: 'when mentioned' })
+    const sqlDefault = new Builder(schema)
     it('Produces [SELECT "col1" FROM "testTable" ORDER BY "col1" NULLS LAST;] option(always)', () => {
       const actual = sqlAlways
         .select(column1)
@@ -215,9 +215,9 @@ describe('test Options', () => {
   })
 
   describe('test SelectItems AS Option', () => {
-    const sqlAlways = new Builder(db, { addAsBeforeColumnAlias: 'always' })
-    const sqlNever = new Builder(db, { addAsBeforeColumnAlias: 'never' })
-    const sqlDefault = new Builder(db)
+    const sqlAlways = new Builder(schema, { addAsBeforeColumnAlias: 'always' })
+    const sqlNever = new Builder(schema, { addAsBeforeColumnAlias: 'never' })
+    const sqlDefault = new Builder(schema)
     it('Produces [SELECT "col1" AS "C1" FROM "testTable";] option(always)', () => {
       const actual = sqlAlways
         .select(column1.as('C1'))