+++ test/error.test.ts
@@ -2,7 +2,7 @@ import {
   Builder,
   BooleanColumn,
   ColumnNotFoundError,
-  Database,
+  Schema,
   e,
   NumberColumn,
   Table,
@@ -36,8 +36,8 @@ describe('Throw desired Errors', () => {
     'testTable',
     [column1, column2, column3, column4, column5, column6, column7, column8],
   )
-  const db = new Database([table], 1)
-  const sql = new Builder(db)
+  const schema = new Schema([table])
+  const sql = new Builder(schema)
 
   it('Throws error when add invalid operator', () => {
     function actual() {