+++ test/orderByStep.test.ts
@@ -1,5 +1,5 @@
 import {
-  Database,
+  Schema,
   Table,
   BooleanColumn,
   NumberColumn,
@@ -28,8 +28,8 @@ describe('test orderBy Step', () => {
     'testTable',
     [column1, column2, column3, column4, column5, column6, column7, column8],
   )
-  const db = new Database([table], 1)
-  const sql = new Builder(db)
+  const schema = new Schema([table])
+  const sql = new Builder(schema)
 
   it('Produces [SELECT * FROM "testTable" ORDER BY "col1", "col2";]', () => {
     const actual = sql